Revert "Revert's the healthchare_visits.py rename for review purposes"
This reverts commit 8709f94583a8718fe88f298b86ded73cc15d4d99.
